Update Find_Stream.py
The previous version was making entries of streams every time it matches the criteria, but this version would not do that. Instead, it finds the identical ones and just increments the channels, suggesting that the stream has x number of channels of the specific type of codec provided by the user.