Added the tsplayout-auto script
This script enables playing streams on the device itself, without any use of streamXpress or any other software that basically runs or streams compressed streams on external device. It also has a .service file that essentially makes the script run as a service whenever the server is powered and stays up an running.