Backup Automation
This python script basically grabs all the essential data required from a server and saves it onto a local machine, through which we have established a connection to the server in the first place. 

This can help in situations if and when a server crashes, as long as we have access to network, we can still grab the important stuff, so that when we re-image it, it could be useful to just re-apply the backup.